stream_edit: Move set_stream_property to new stream_settings_api module.
